icon-service/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.0.1+cpu
torchvision==0.15.2+cpu
diffusers==0.21.4
transformers==4.31.0
accelerate==0.21.0
safetensors==0.3.1
numpy==1.24.3
pillow==10.0.0
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
huggingface-hub==0.16.4
datasets==2.14.7